requirements.txt
tortoise-orm[asyncpg]
aerich
sanic[ext]
pydantic
tqdm 
requests